Box Layout and Icon Image
The FlowLayout automatically wraps the components and sometimes we don't want this to happen, so, I changed the main panel to BoxLayout to disable the automatic wrapping of components. I added several panels to group the buttons accordingly.

I added also an icon image to override the default icon of java programs. The icon is my own icon, all rights are reserved.